Move testQueryChangeAllParameters to specific test
