+++ tests/Integration/ConsumeServiceTestCase.php
@@ -8,22 +8,14 @@ use GuzzleHttp\Client as GuzzleClient;
 use GuzzleHttp\RequestOptions;
 use PhpCfdi\SatWsDescargaMasiva\Service;
 use PhpCfdi\SatWsDescargaMasiva\Services\Query\QueryParameters;
-use PhpCfdi\SatWsDescargaMasiva\Shared\CfdiComplemento;
-use PhpCfdi\SatWsDescargaMasiva\Shared\DateTimePeriod;
-use PhpCfdi\SatWsDescargaMasiva\Shared\DocumentStatus;
-use PhpCfdi\SatWsDescargaMasiva\Shared\DocumentType;
-use PhpCfdi\SatWsDescargaMasiva\Shared\DownloadType;
-use PhpCfdi\SatWsDescargaMasiva\Shared\RequestType;
-use PhpCfdi\SatWsDescargaMasiva\Shared\RfcMatch;
-use PhpCfdi\SatWsDescargaMasiva\Shared\RfcOnBehalf;
-use PhpCfdi\SatWsDescargaMasiva\Shared\Uuid;
+use PhpCfdi\SatWsDescargaMasiva\Shared\ServiceEndpoints;
 use PhpCfdi\SatWsDescargaMasiva\Tests\TestCase;
 use PhpCfdi\SatWsDescargaMasiva\WebClient\GuzzleWebClient;
 use PhpCfdi\SatWsDescargaMasiva\WebClient\WebClientInterface;
 
 abstract class ConsumeServiceTestCase extends TestCase
 {
-    abstract protected function createService(): Service;
+    abstract protected function getServiceEndpoints(): ServiceEndpoints;
 
     protected function createWebClient(): WebClientInterface
     {
@@ -32,6 +24,13 @@ abstract class ConsumeServiceTestCase extends TestCase
             RequestOptions::TIMEOUT => 30,
         ]);
         return new GuzzleWebClient($guzzleClient);
+    }
+
+    protected function createService(): Service
+    {
+        $requestBuilder = $this->createFielRequestBuilderUsingTestingFiles();
+        $webClient = $this->createWebClient();
+        return new Service($requestBuilder, $webClient, null, $this->getServiceEndpoints());
     }
 
     public function testAuthentication(): void
@@ -46,31 +45,6 @@ abstract class ConsumeServiceTestCase extends TestCase
         $service = $this->createService();
 
         $parameters = QueryParameters::create();
-
-        $result = $service->query($parameters);
-        $this->assertSame(
-            305,
-            $result->getStatus()->getCode(),
-            'Expected to receive a 305 - Certificado Inválido from SAT since FIEL is for testing'
-        );
-    }
-
-    public function testQueryChangeAllParameters(): void
-    {
-        $service = $this->createService();
-
-        $parameters = QueryParameters::create()
-            ->withPeriod(DateTimePeriod::createFromValues('2019-01-01 00:00:00', '2019-01-01 00:04:00'))
-            ->withDownloadType(DownloadType::received())
-            ->withRequestType(RequestType::cfdi())
-            ->withDocumentType(DocumentType::nomina())
-            /** @todo uncomment this line when SAT service is working */
-            // ->withComplement(CfdiComplemento::leyendasFiscales10())
-            ->withDocumentStatus(DocumentStatus::active())
-            ->withUuid(Uuid::create('96623061-61fe-49de-b298-c7156476aa8b'))
-            ->withRfcOnBehalf(RfcOnBehalf::create('XXX01010199A'))
-            ->withRfcMatch(RfcMatch::create('AAA010101AAA'))
-        ;
 
         $result = $service->query($parameters);
         $this->assertSame(
